move /bank award to team
