+++ modules/team.js
@@ -3,6 +3,7 @@ const Augur = require("augurbot-ts");
 const Discord = require("discord.js");
 const u = require("../utils/utils");
 const c = require("../utils/modCommon");
+const { bankVars } = require("./bank");
 
 /**
  * @typedef {"team"|"mod"|"mgr"|"mgmt"} perms
@@ -25,6 +26,80 @@ async function slashTeamRoleGive(int, give = true) {
   } catch (error) { u.errorHandler(error, int); }
 }
 
+/** @param {Discord.GuildMember} member */
+function getHouseInfo(member) {
+  const houseInfo = new Map([
+    [u.sf.roles.housebb, { name: "Brightbeam", color: 0x00a1da }],
+    [u.sf.roles.housefb, { name: "Freshbeast", color: 0xfdd023 }],
+    [u.sf.roles.housesc, { name: "Starcamp", color: 0xe32736 }]
+  ]);
+
+  for (const [k, v] of houseInfo) {
+    if (member.roles.cache.has(k)) return v;
+  }
+  return { name: "Unsorted", color: 0x402a37 };
+}
+
+/** @param {Augur.GuildInteraction<"CommandSlash">} interaction */
+async function slashTeamBankAward(interaction) {
+  try {
+    const { ember, gb, limit } = bankVars;
+    const giver = interaction.member;
+    const recipient = interaction.options.getMember("user");
+    const reason = interaction.options.getString("reason") || "Astounding feats of courage, wisdom, and heart";
+    let value = interaction.options.getInteger("amount", true);
+    if (!recipient) return interaction.reply({ content: "I couldn't find that user!", ephemeral: true });
+
+    let reply = "";
+
+    if (recipient.id === giver.id) {
+      reply = `You can't award ***yourself*** ${ember}, silly.`;
+    } else if (recipient.id !== interaction.client.user.id && recipient.user.bot) {
+      reply = `Bots don't really have a use for awarded ${ember}.`;
+    } else if (value === 0) {
+      reply = "You can't award ***nothing***.";
+    }
+
+    if (reply) return interaction.reply({ content: reply, ephemeral: true });
+
+    value = value < 0 ? Math.max(value, -1 * limit.ember) : Math.min(value, limit.ember);
+
+    const award = {
+      currency: "em",
+      discordId: recipient.id,
+      description: `From ${giver.displayName} (House Points): ${reason}`,
+      value,
+      giver: giver.id,
+      hp: true
+    };
+
+    const receipt = await u.db.bank.addCurrency(award);
+    const balance = await u.db.bank.getBalance(recipient.id);
+    const str = (/** @type {string} */ m) => value > 0 ? `awarded ${m} ${ember}${receipt.value}` : `docked ${ember}${-receipt.value} from ${m}`;
+    let embed = u.embed({ author: interaction.client.user })
+      .addFields(
+        { name:"Reason", value: reason },
+        { name: "Your New Balance", value: `${gb}${balance.gb}\n${ember}${balance.em}` }
+      )
+      .setDescription(`${u.escapeText(giver.displayName)} just ${str("you")}! This counts toward your House's Points.`);
+
+    await interaction.reply(`Successfully ${str(recipient.displayName)} for ${reason}. This counts towards their House's Points.`);
+    recipient.send({ embeds: [embed] }).catch(() => interaction.followUp({ content: `I wasn't able to alert ${recipient} about the award. Please do so yourself.`, ephemeral: true }));
+    u.clean(interaction, 60000);
+
+    const house = getHouseInfo(recipient);
+
+    embed = u.embed({ author: recipient })
+      .setColor(house.color)
+      .addFields(
+        { name: "House", value: house.name },
+        { name: "Reason", value: reason }
+      )
+      .setDescription(`**${giver}** ${str(recipient.toString())}`);
+    interaction.client.getTextChannel(u.sf.channels.mopbucketawards)?.send({ embeds: [embed] });
+  } catch (e) { u.errorHandler(e, interaction); }
+}
+
 
 const Module = new Augur.Module()
 .addInteraction({ id: u.sf.commands.slashTeam,
@@ -34,6 +109,7 @@ const Module = new Augur.Module()
     switch (int.options.getSubcommand()) {
       case "give": return slashTeamRoleGive(int, true);
       case "take": return slashTeamRoleGive(int, false);
+      case "award": return slashTeamBankAward(int);
       default: return u.errorHandler(new Error("Unhandled Subcommand"), int);
     }
   },
