+++ registry/slashTeam.js
@@ -1,6 +1,8 @@
 // @ts-check
 const u = require("./regUtils");
 
+// ROLE COMMANDS
+
 const role = (action = "add") => new u.string()
   .setName(action === 'equip' ? "color" : "role")
   .setDescription(`The role to ${action}`)
@@ -29,6 +31,33 @@ const take = new u.sub()
 )
 .addStringOption(role("take"));
 
+
+// BANK COMMANDS
+
+const award = new u.sub()
+  .setName("award")
+  .setDescription("Award ember to a member for the house cup.")
+  .addUserOption(
+    new u.user()
+      .setName("user")
+      .setDescription("Who do you want to award?")
+      .setRequired(true)
+  )
+  .addIntegerOption(
+    new u.int()
+      .setName("amount")
+      .setDescription("How many ember do you want to give them?")
+      .setRequired(true)
+      .setMinValue(1)
+      .setMaxValue(10000)
+  )
+  .addStringOption(
+    new u.string()
+      .setName("reason")
+      .setDescription("But... why?")
+      .setRequired(false)
+  );
+
 module.exports = new u.cmd()
   .setName("team")
   .setDescription("Do team stuff. idk.")
@@ -38,6 +67,12 @@ module.exports = new u.cmd()
       .setDescription("Manage roles for a user.")
       .addSubcommand(give)
       .addSubcommand(take)
+  )
+  .addSubcommandGroup(
+    new u.subGroup()
+      .setName("bank")
+      .setDescription("Interact with currency")
+      .addSubcommand(award)
   )
   .setDMPermission(false)
   .setDefaultMemberPermissions(u.privateCommand)
